Fix DSE tests
